refactor: consolidate to msgs.go
